committing changes to servingsCounterPage, before restructuring files and isolating components for more 'React'ivity. Should have done it before now.
