+++ client/src/components/circles/LogO.js
@@ -1,11 +1,16 @@
 import React from 'react';
 import './circle.css';
+import { Link } from '@mui/material';
 
 export default function LogO() {
     return (
+      <>
+      <Link to ="/log">
       <div className="circle">
         <div className="circleText2">Log</div>
       </div>
+      </Link>
+      </>
     );
     }
