--- client/src/components/circles/SleepO.js
@@ -1,10 +1,0 @@
-import React from "react";
-import "./circle.css"
-
-export default function SleepO() {
-    return (
-      <div className="circle">
-        <div className="circleText2">Sleep</div>
-      </div>
-    );
-    }
